requirements.txt
fastapi
uvicorn[standard]
streamlit
python-dotenv
qdrant-client
langchain
langchain-community
langchain-text-splitters
sentence-transformers
langchain-qdrant
unstructured
google-generativeai
langchain-google-genai
requests
tavily-python
